fix conversion error

+++ tests/test_observation_models.py
@@ -214,7 +214,7 @@ class TestPoissonObservations:
         """
         _, y, model, _, firing_rate = poissonGLM_model_instantiation
         ll_model = model.observation_model.log_likelihood(
-            y, firing_rate, aggregate_sample_scores=lambda x: x
+            y.astype(float), firing_rate, aggregate_sample_scores=lambda x: x
         )
         ll_scipy = sts.poisson(firing_rate).logpmf(y)
         if not np.allclose(ll_model, ll_scipy):
